+++ Evergreen/Apps/Get-OperaBrowser.ps1
@@ -1,4 +1,4 @@
-Function Get-OperaBrowser {
+function Get-OperaBrowser {
     <#
         .SYNOPSIS
             Returns the available Opera Browser versions and download URIs.
@@ -17,22 +17,28 @@ Function Get-OperaBrowser {
         $res = (Get-FunctionResource -AppName ("$($MyInvocation.MyCommand)".Split("-"))[1])
     )
 
-    $Update = Invoke-RestMethodWrapper -Uri $res.Get.Update.Uri
-    Write-Verbose -Message "$($MyInvocation.MyCommand): checking property: $($res.Get.Update.Property)."
-    Write-Verbose -Message "$($MyInvocation.MyCommand): found version: $($Update.current_version)"
-    If ($Null -ne $Update.($res.Get.Update.Property)) {
+    foreach ($Channel in $res.Get.Update.Channels) {
+        $Update = Invoke-RestMethodWrapper -Uri $res.Get.Update.Uri[$Channel]
 
-        # Step through each installer type
-        ForEach ($item in $res.Get.Download.Uri.GetEnumerator()) {
+        if ($Null -ne $Update.($res.Get.Update.Property)) {
+            Write-Verbose -Message "$($MyInvocation.MyCommand): checking property: $($res.Get.Update.Property)."
+            Write-Verbose -Message "$($MyInvocation.MyCommand): found version: $($Update.($res.Get.Update.Property))"
 
-            # Build the output object; Output object to the pipeline
-            $PSObject = [PSCustomObject] @{
-                Version      = $Update.($res.Get.Update.Property)
-                Architecture = $item.Name
-                Type         = Get-FileType -File $res.Get.Download.Uri[$item.Key]
-                URI          = $res.Get.Download.Uri[$item.Key] -replace $res.Get.Download.ReplaceText, $Update.($res.Get.Update.Property)
+            # Step through each installer type
+            foreach ($Architecture in $res.Get.Download.Architectures) {
+
+                # Build the output object; Output object to the pipeline
+                $Url = $res.Get.Download.Uri[$Channel] -replace $res.Get.Download.ReplaceText, $Update.($res.Get.Update.Property) `
+                    -replace "#architecture", $res.Get.Download.Architecture[$Architecture]
+                $PSObject = [PSCustomObject] @{
+                    Version      = $Update.($res.Get.Update.Property)
+                    Channel      = $Channel
+                    Architecture = $Architecture
+                    Type         = Get-FileType -File $Url
+                    URI          = $Url
+                }
+                Write-Output -InputObject $PSObject
             }
-            Write-Output -InputObject $PSObject
         }
     }
 }